+++ src/converse-muc-views.js
@@ -1400,6 +1400,20 @@
                     return;
                 },
 
+                getNotificationWithMessage (message) {
+                    let el = this.content.lastElementChild;
+                    while (!_.isNil(el)) {
+                        const data = _.get(el, 'dataset', {});
+                        if (!_.includes(_.get(el, 'classList', []), 'chat-info')) {
+                            return;
+                        }
+                        if (el.textContent === message) {
+                            return el;
+                        }
+                        el = el.previousElementSibling;
+                    }
+                },
+
                 parseXUserElement (x, stanza, is_self) {
                     /* Parse the passed-in <x xmlns='http://jabber.org/protocol/muc#user'>
                      * element and construct a map containing relevant
@@ -1409,7 +1423,10 @@
                     const statuses = x.querySelectorAll('status');
                     const mapper = _.partial(this.getMessageFromStatus, _, stanza, is_self);
                     const notification = {};
-                    const messages = _.reject(_.map(statuses, mapper), _.isUndefined);
+                    const messages = _.reject(
+                        _.reject(_.map(statuses, mapper), _.isUndefined),
+                        message => this.getNotificationWithMessage(message)
+                    );
                     if (messages.length) {
                         notification.messages = messages;
                     }